requirements.txt
Pillow
h5py
numpy
scikit-image
tensorboard
torchvision
pytorch
xmltodict
